Implement parallel routing for members (#204)
* implement parallel routing on invite member modal

* cleanup

* parallel routing for deactivate and reactivate members
