+++ src/app/(organization)/members/[member]/deactivate/page.tsx
@@ -1,91 +1,15 @@
 'use client';
-import { Button, Form, Modal } from '@holaplex/ui-library-react';
-import Card from '../../../../../components/Card';
-import Typography, { Size } from '../../../../../components/Typography';
-import { useRouter } from 'next/navigation';
-import { Icon } from '../../../../../components/Icon';
-import { Member, DeactivateMemberInput } from '../../../../../graphql.types';
-import { useMutation } from '@apollo/client';
-import { DeactivateMember } from '../../../../../mutations/member.graphql';
-import { GetOrganizationMembers } from '../../../../../queries/organization.graphql';
-import { useOrganization } from '../../../../../hooks/useOrganization';
+import React from 'react';
+import DeactivateMember from '../../DeactivateMember';
 
-interface DeleteMemberData {
-  deleteMember: Member;
-}
-
-interface DeleteMemberVars {
-  input: DeactivateMemberInput;
-}
-
-interface DeleteMemberProps {
+interface DeactivateMemberPageProps {
   params: { member: string };
 }
 
-export default function MemberDeletePage({ params: { member } }: DeleteMemberProps) {
-  const { organization } = useOrganization();
-  const router = useRouter();
-
-  const onClose = () => {
-    router.push('/members');
-  };
-
-  const [deleteMember, { loading }] = useMutation<DeleteMemberData, DeleteMemberVars>(
-    DeactivateMember,
-    {
-      variables: {
-        input: {
-          id: member,
-        },
-      },
-      awaitRefetchQueries: true,
-      refetchQueries: [
-        { query: GetOrganizationMembers, variables: { organization: organization?.id } },
-      ],
-    }
-  );
-
-  const onDelete = () => {
-    deleteMember({
-      variables: {
-        input: {
-          id: member,
-        },
-      },
-      onCompleted: () => {
-        router.push('/members');
-      },
-    });
-  };
-
+export default function DeactivateMemberPage({ params: { member } }: DeactivateMemberPageProps) {
   return (
-    <Modal open={true} setOpen={onClose}>
-      <Card className="w-[400px]">
-        <Typography.Header size={Size.H2}>Do you really want to deactivate the member?</Typography.Header>
-        <Typography.Header size={Size.H3}>This action can be reversed.</Typography.Header>
-
-        <Form className="flex flex-col mt-5">
-          <div className="flex gap-2 rounded-md bg-stone-800 p-3">
-            <Icon.Info stroke="stroke-gray-400" />
-            <span className="text-xs text-gray-400 font-medium text-left">
-              The member will be deprived of access to the project
-            </span>
-          </div>
-          <Button
-            htmlType="submit"
-            className="w-full mt-5"
-            variant="failure"
-            onClick={onDelete}
-            disabled={loading}
-            loading={loading}
-          >
-            Deactivate member
-          </Button>
-          <Button variant="secondary" className="w-full mt-5" onClick={onClose} disabled={loading}>
-            Cancel
-          </Button>
-        </Form>
-      </Card>
-    </Modal>
+    <div className="h-screen flex items-center justify-center">
+      <DeactivateMember member={member} />
+    </div>
   );
 }
